+++ Console/Command/CatalogSyncAttributesSaveCommand.php
@@ -49,8 +49,8 @@ class CatalogSyncAttributesSaveCommand extends BaseCommand
      */
     public function configure()
     {
-        $this->setName('flow:flow-connector:catalog-sync-attributes-save')
-            ->setDescription('Saves product attributes needed for the integration in flow.io');
+        $this->setName('flow:connector:catalog-attributes-save')
+            ->setDescription('Save product attributes needed for catalog integration to Flow.');
     }
 
     /**