+++ Console/Command/WebhookUpdateSettingsCommand.php
@@ -44,8 +44,8 @@ class WebhookUpdateSettingsCommand extends BaseCommand
 
     public function configure()
     {
-        $this->setName('flow:flow-connector:webhook-update-settings')
-            ->setDescription('Register webhooks with Flow.');
+        $this->setName('flow:connector:webhook-update')
+            ->setDescription('Create secret for webhook payload verification.');
     }
 
     /**